Se adiciona variables de entorno y se separan los endpoints de index.js a app.js
