+++ src/controllers/employees.controllers.js
@@ -1,4 +1,4 @@
-import { pool } from "../db.js"
+import { pool } from '../db.js'
 
 
 // Obtener un solo dato
