+++ src/index.js
@@ -1,20 +1,7 @@
-import express from 'express';
-import indexRoutes from './routes/index.routes.js'
-import employeesRoutes from './routes/employees.routes.js'
+import app from './app.js'
+import { PORT } from './config.js'
 
-const app = express()
 
-app.use(express.json())
+app.listen(PORT)
+console.log('Server running on port',PORT)
 
-app.use('/v1',indexRoutes)
-app.use('/v1',employeesRoutes)
-
-// mensaje para cuando no se encuentre el endpoint
-app.use((req, res, next) => {
-    res.status(404).json({
-        message:'Not Found'
-    })
-})
-
-app.listen(3000)
-console.log('server running on port 3000')
