+++ src/routes/employees.routes.js
@@ -1,13 +1,14 @@
 import { Router } from 'express'
 import {getEmployee, getEmployees, createEmployee, updateEmployee, deleteEmployee} from '../controllers/employees.controllers.js'
+
+
 const router  = Router()
 
-
-router.get('/employes', getEmployees)
-router.get('/employes/:id', getEmployee)
-router.post('/employes', createEmployee )
-router.patch('/employes/:id', updateEmployee)
-router.delete('/employes/:id', deleteEmployee)
+router.get('/employees', getEmployees)
+router.get('/employees/:id', getEmployee)
+router.post('/employees', createEmployee )
+router.patch('/employees/:id', updateEmployee)
+router.delete('/employees/:id', deleteEmployee)
 
 
 export default router
